feat: add exception and handle exception (#7)
* Handle exception currencyapi

* Handle limit amount exception

* Handle connect to conditional bank exception

* Handle transactions exception

* Change common connect exception on custom for currencyapi

* Realize handle transactions exceptions

+++ thrift-bank/src/main/java/efr/iv/igr/thriftbank/service/impl/TransactionServiceImpl.java
@@ -1,6 +1,8 @@
 package efr.iv.igr.thriftbank.service.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import efr.iv.igr.thriftbank.exception.InvalidAmountException;
+import efr.iv.igr.thriftbank.exception.SimilarIndetifierException;
 import efr.iv.igr.thriftbank.mapper.TransactionMapper;
 import efr.iv.igr.thriftbank.model.entity.Transaction;
 import efr.iv.igr.thriftbank.model.request.TransactionRequest;
@@ -11,6 +13,7 @@ import efr.iv.igr.thriftbank.service.ITransactionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -31,7 +34,15 @@ public class TransactionServiceImpl implements ITransactionService {
     }
 
     @Override
-    public TransactionResponse createTransaction(TransactionRequest transactionRequest) {
+    public TransactionResponse createTransaction(TransactionRequest transactionRequest) throws InvalidAmountException, SimilarIndetifierException {
+        if (transactionRequest.getAmount().compareTo(BigDecimal.valueOf(100)) <= 0) {
+            throw new InvalidAmountException("Amount must be bigger than 100");
+        }
+
+        if (transactionRequest.getAccountFrom().compareTo(transactionRequest.getAccountTo()) == 0) {
+            throw new SimilarIndetifierException("You can't transfer money to yourself");
+        }
+
         cacheTransactionService.deleteCacheData();
         Transaction transaction = transactionMapper.toEntity(transactionRequest);
         transaction.setDate(Instant.now());
